Fix issue with tests on travis.
There were several issues with running on Travis.  (1) bcrypt is evil.  A change in how it is deployed required that we import it once in python with sudo priveleges.  This resulted in a very cryptic 'File Not Found' error when trying to register a user.  (2) The Romanesco package was still pointing to arborworkflows, not the Kitware repo.

I changed some of the test code to match what was done in girder.  I added the screenshot output method.  I changed how the web client test is called so that the test output is interleaved with the stdout and stderr of the process (this makes it easier to see what order things occurred in).

I changed Javascript Code Style rules to match what is being done; the defaults on my system throw a number of warnings.