Fix the travis build.
Primarily, fix the tangelo version to 0.7.0 until we are ready to update the tests.

Change how tangelo is called so that we can log error output.

Add vmstat to tell if travis is having a memory issue.  Added a quick test that errors out the travis build before doing much if the memory issue is happening.