Fixes #150. Pass auth token to romanesco
This is the minimal change required to fix the bug where private
data could not be downloaded from girder in the romanesco worker.
Many places in the code expected the user model to contain a
"token" field, which it did not. So the easiest thing was just to
set that field on the current user model at startup or login time.